Created lendor_registration_form.dashboard.top_up_amount, lendor_registration_form.dashboard.view_available_balance, lendor_registration_form.dashboard.view_borrower_loan_request, lendor_registration_form.dashboard.view_borrower_loan_request.Borr_loan_request and 2 more; Deleted lendor_registration_form.dashboard.avlbal, lendor_registration_form.dashboard.rta, lendor_registration_form.dashboard.vblr, lendor_registration_form.dashboard.vblr.Borr_loan_request and 2 more
