+++ .anvil_editor.yaml
@@ -23,11 +23,9 @@ unique_ids:
     wallet.wallet: '1703489007387496965863254.889'
     lendor_registration_form.Lender_reg_form_2: '1699426727342910603650280.5131'
     borrower_registration_form.dashboard.edit_profile: '1698653709394569134493869.6183'
-    lendor_registration_form.dashboard.avlbal: '1699960014867483301520306.9306'
     admin.dashboard.manage_reports: '1699528696451136679061767.8162'
     admin.user_issue.user_buganalysis.RowTemplate17: '1701189107287981500115093.1753'
     lendor_registration_form.dashboard.lender_view_loans.RowTemplate21: '1703140075771292076881029.6881'
-    lendor_registration_form.dashboard.vblr.RowTemplate10: '1700542200026541460267637.76886'
     lendor_registration_form.Lender_reg_form_4: '1700038033302654577889349.2847'
     borrower_registration_form.star_1_borrower_registration_form_begin_8: '1700161372617325052945266.5503'
     admin.dashboard.manage_settings.manage_facts: '1699529264900376484216542.50616'
@@ -49,7 +47,6 @@ unique_ids:
     lendor_registration_form.dashboard.today_dues.View_Details: '170356566713748309638072.5887'
     admin.dashboard.lenders.view_profile_copy.edit_form_copy: '1700738442974165215211725.12378'
     lendor_registration_form.Toggleswitch: '1701335189405602120940903.774'
-    lendor_registration_form.dashboard.vblr.Borr_loan_request: '1700542763128106365112272.9714'
     borrower_registration_form.dashboard.foreclosure_request: '1699009486604204061292370.0036'
     admin.dashboard.admin_teams.view_profile.update_form: '169952839624818688697299.6676'
     lendor_registration_form.dashboard.today_dues.ItemTemplate1: '1703564885286922717800272.7809'
@@ -62,6 +59,7 @@ unique_ids:
     wallet.wallet.all_transaction.RowTemplate29: '1704884931461492157927200.2286'
     lendor_registration_form.dashboard.edit_profile: '1701093412433606982932624.6844'
     admin.dashboard.revenue_share: '1699529452988469665100106.9467'
+    lendor_registration_form.dashboard.view_borrower_loan_request: '1699960229158369964369354.56226'
     admin.dashboard.loan_management.loan_disbursed.view_profile_6: '1704878672712960872612311.6324'
     bank_users.main_form.career_main_form: '1700563797991728540622095.9799'
     admin.dashboard.loan_management.closed_loans.RowTemplate14: '1701079863491503664452804.4547'
@@ -82,6 +80,7 @@ unique_ids:
     bank_users.user_form.user_form_copy: '1700712538416515803210372.55963'
     admin.dashboard.borrowers.view_profile.update_form: '1699967990279922551849592.2898'
     borrower_registration_form.star_1_borrower_registration_form_begin_3.star_1_borrower_registration_form_begin_3c: '1699373440164759513719499.4375'
+    lendor_registration_form.dashboard.view_borrower_loan_request.RowTemplate10: '1700542200026541460267637.76886'
     borrower_registration_form.dashboard.dashboard_about: '1701024263832190033852688.87518'
     borrower_registration_form.dashboard.borrower_profile: '1698653519986110665503267.72595'
     admin.dashboard.manage_products.view_categories.ItemTemplate2: '1704371055344928457413512.8632'
@@ -92,7 +91,6 @@ unique_ids:
     admin.dashboard.loan_management.rejected_loans: '1699528958889169021462649.09222'
     admin.dashboard.loan_management.rejected_loans.view_profile_2: '1700812221243295088415002.8101'
     admin.dashboard.loan_management.approved_loans.view_profile: '1700740546893987573171578.3491'
-    lendor_registration_form.dashboard.rta: '1699957883953144866963289.06934'
     ? borrower_registration_form.star_1_borrower_registration_form_begin_3.star_1_borrower_registration_form_begin_3b_business_3.star_1_borrower_registration_form_begin_3b_business_2
     : '1700570068782198167271670.87088'
     admin.dashboard.loan_management.loan_disbursed: '1701159817062849579236397.6018'
@@ -101,15 +99,16 @@ unique_ids:
     admin.dashboard.loan_management.rejected_loans.RowTemplate8: '1700812073713245303231724.8567'
     bank_users.main_form.about_main_form: '1700562950795405574544713.0248'
     borrower_registration_form.dashboard.foreclosure_request.borrower_foreclosure: '1700465774321790534045777.5126'
-    lendor_registration_form.dashboard.vblr: '1699960229158369964369354.56226'
     lendor_registration_form.Lender_reg_individual_form_1: '169942903909778939193987.11401'
     admin.dashboard.admin_teams.view_profile.edit_form: '1699528319288104807163991.37613'
     lendor_registration_form.dashboard.lender_view_loans.RowTemplate20: '1703139972848360547706878.1784'
     lendor_registration_form.dashboard.vlfr.foreclose_details: '1700454637827919052628054.0963'
     lendor_registration_form.Lender_reg_form_3: '1699427114774355019321073.1847'
     admin.user_issue.user_viewprofile: '1701187753190491873666760.6118'
+    lendor_registration_form.dashboard.top_up_amount: '1699957883953144866963289.06934'
     lendor_registration_form.Lender_reg_individual_form_3: '1699429299471944361321501.9009'
     borrower_registration_form.dashboard.application_tracker.view_profile: '1701144654092613273531448.6368'
+    lendor_registration_form.dashboard.view_borrower_loan_request.Borr_loan_request: '1700542763128106365112272.9714'
     lendor_registration_form.Lender_reg_bankdirect_bank_form_2: '1699429556412724511736888.3689'
     admin.dashboard.manage_products.view_product: '1700713112442254456792559.71744'
     admin.dashboard.loan_management.closed_loans: '1699528886441906086843573.9337'
@@ -119,7 +118,6 @@ unique_ids:
     borrower_registration_form.star_1_borrower_registration_form_begin_3.star_1_borrower_registration_form_begin_3b_business_3: '1700126739765126145778650.8795'
     borrower_registration_form.dashboard.foreclosure_request.RowTemplate6: '1700745206852325214857298.85956'
     borrower_registration_form.dashboard.foreclosure_request.borrower_foreclosure.foreclose: '1700730776806358375111007.35254'
-    lendor_registration_form.dashboard.vblr.Borr_loan_request.RowTemplate11: '1700543062684858008325535.9921'
     admin.dashboard.manage_products.manage_producs1: '1700713086879771707093045.5188'
     admin.dashboard.manage_cms: '1699528617179453409715572.5991'
     lendor_registration_form.Lender_reg_Institutional_form_3: '1699430137541808903451295.2183'
@@ -142,6 +140,7 @@ unique_ids:
     admin.dashboard.borrowers.view_profile: '1699967781060728634256553.5089'
     admin.dashboard.borrowers.view_profile.edit_form: '169996780467215868230296.983032'
     lendor_registration_form.dashboard.lender_view_loans.RowTemplate23: '1703141182461211155517631.55914'
+    lendor_registration_form.dashboard.view_available_balance: '1699960014867483301520306.9306'
     ? borrower_registration_form.star_1_borrower_registration_form_begin_3.star_1_borrower_registration_form_begin_3b_business_3.star_1_borrower_registration_form_begin_3b_business_5
     : '17005720020403876226355.3249187'
     borrower_registration_form.dashboard.application_tracker: '1698927021026141533584098.39868'
@@ -182,6 +181,7 @@ unique_ids:
     borrower_registration_form.dashboard.application_tracker.RowTemplate26: '1704699099052678345492761.1499'
     lendor_registration_form.dashboard.lender_view_loans.RowTemplate22: '1703141013007184309847389.77017'
     admin.dashboard.loan_management.closed_loans.view_profile_5: '1701079991010941258272977.0447'
+    lendor_registration_form.dashboard.view_borrower_loan_request.Borr_loan_request.RowTemplate11: '1700543062684858008325535.9921'
     lendor_registration_form.dashboard.vlfr.RowTemplate30: '1704950843840196993818177.19797'
     wallet.wallet.all_transaction: '1704884604578852863514289.3086'
     lendor_registration_form.dashboard.vlfr.RowTemplate7: '1700215849908648465910279.8993'
