+++ client_code/lendor_registration_form/dashboard/view_available_balance/form_template.yaml
