+++ client_code/lendor_registration_form/dashboard/view_borrower_loan_request/Borr_loan_request/__init__.py
@@ -49,7 +49,7 @@ class Borr_loan_request(Borr_loan_requestTemplate):
                 # Fetch additional details from the 'loan_details' table
                 try:
                     #loan_details = app_tables.loan_details.get(loan_id=int(selected_row['loan_id']))
-                    loan_details = app_tables.fin_loan_details.get(loan_id=str(selected_row['loan_id']))
+                    loan_details = app_tables.loan_details.get(loan_id=str(selected_row['loan_id']))
                     if loan_details is not None:
                         # Assuming 'interest_rate' and 'min_amount' are valid column names in the 'loan_details' table
                         interest_rate = loan_details['interest_rate']
@@ -205,11 +205,8 @@ class Borr_loan_request(Borr_loan_requestTemplate):
 
         # Check the signal and perform actions accordingly
         if signal == "insufficient_balance":
-            alert("Warning: Your account balance is insufficient. Please deposit amount into your wallet. If not done within the next 2 minutes, the opportunity may be lost")
-            # open_form("wallet.wallet")
-        elif signal == "Time_out":
-            alert("The designated time has passed. The loan has moved to the 'Lost Opportunities' status.")
-            open_form("wallet.wallet") 
+            alert("Insufficient balance. Please add the amount in the wallet. If not added within 30 min, it will go to lost opportunities.")
+            open_form("wallet.wallet")
         elif signal == "pay_to_borrower":
             alert("Pay to Borrower")
             open_form("wallet.wallet")
